+++ HISTORY.rst
@@ -15,6 +15,7 @@ tnefparse 1.4.0 (unreleased)
 - add expirimental support for parsing embedded message objects (jrideout)
 - zipped output now uses long filename when possible (jrideout)
 - run GitHub Actions CI for pull requests and once a week (eumiro)
+- use pathlib instead of os.path (eumiro)
 
 tnefparse 1.3.1 (2020-09-30)
 =============================